[Leverage AAVE for deposit]
A smart contract that can be used as a deposit proxy for sharedStake.  
The idea is to leverage AAVE to allow users to go long or short on ETH when staking for rewards.  

E.g. 
- A fiat heavy or short ETH investor may be worried about ETH volatility and prefer to deposit StableCoins as collateral into the sharedStake yield generating platform. In this case, this proxy contract will deposit stables into AAVE, withdraw ETH, and deposit that to sharedStake

- Another case is an investor that is long ETH. In this case instead of locking up ETH directly into sharedStake, the ETH is invested in AAVE, and up-to a health factor of 1.5,  stablecoins (DAI) is withdrawn. The DAI is swapped via uniswap to ETH which is then invested. This allows the investor to gain liquidity from any ETH price increases. 

For both cases, any remaining credit is delegated back to the user.   
Currently the tokens remain in the smart contract.  
TODO: 
- Add send to sharedStake, and return minted vETH2 to user
- Test on mainnet